Fix collection document ref
